Show all accounts
If there was an account Foo:bar, and all the transactions were for
Foo:bar instead of plain Foo, Foo would never show up in any account
listings. This was problematic because it was expected that bal Foo
would have an entry for all of Foo (which would contain Foo:bar and
Foo:bat).